requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
supabase==2.18.1
google-generativeai==0.8.5
python-multipart==0.0.20
pydantic==2.11.7